refactor: move wizard from Layout to Create
