+++ components/Create/TimeInput.tsx
@@ -1,6 +1,6 @@
-import { useWizardContext } from '../Layout/wizard/Wizard';
-import { FormData, touchDirtyValidate } from '../Layout/wizard/Wizard';
-import styles from '../../components/Layout/wizard/Wizard.module.scss';
+import { useWizardContext } from './wizard/Wizard';
+import { FormData, touchDirtyValidate } from './wizard/Wizard';
+import styles from '../Create/wizard/Wizard.module.scss';
 interface TimeInputProps {
   shortest: string;
   short: string;
